Move some classes into separate files
