+++ Tensile/AsmMemoryInstruction.py
@@ -1,0 +1,84 @@
+################################################################################
+# Copyright 2016-2022 Advanced Micro Devices, Inc. All rights reserved.
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell cop-
+# ies of the Software, and to permit persons to whom the Software is furnished
+# to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IM-
+# PLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+# FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+# COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+# IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNE-
+# CTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+################################################################################
+
+################################################################################
+# Memory Instruction
+################################################################################
+class MemoryInstruction:
+  def __init__(self, name, numAddresses, numOffsets, \
+      offsetMultiplier, blockWidth, formatting):
+    self.name = name
+    self.formatting = formatting
+    self.numAddresses = numAddresses
+    self.numOffsets = numOffsets
+    self.offsetMultiplier = offsetMultiplier
+    self.blockWidth = blockWidth
+    self.numBlocks = 2 if self.numAddresses > 1 or self.numOffsets > 1 else 1
+    self.totalWidth = self.blockWidth * self.numBlocks
+    #in Quad-Cycle
+    if (name == "ds_read_b128"):
+      self.IssueLatency = 2
+    elif (name == "ds_write_b128"):
+      self.IssueLatency = 5
+    elif (name == "ds_write2_b64"):
+      self.IssueLatency = 3
+    elif (name == "ds_write_b64"):
+      self.IssueLatency = 3
+    elif (name == "ds_write2_b32"):
+      self.IssueLatency = 3
+    elif (name == "ds_write_b32"):
+      self.IssueLatency = 2
+    elif (name == "ds_write_u16") :
+      self.IssueLatency = 2
+    else:
+      self.IssueLatency = 1
+    self.endLine = "\n"
+  ########################################
+  # write in assembly format
+  def toString(self, params, comment, nonTemporal=0, highBits=0):
+    name = self.name
+    if highBits:
+      name += "_d16_hi"
+    instStr = "%s %s" % (name, (self.formatting % params) )
+    if nonTemporal%2==1:
+      instStr += " glc"
+    if nonTemporal//2==1:
+      instStr += " slc"
+    line = "%-50s // %s%s" % (instStr, comment, self.endLine)
+    return line
+
+  # Like toString, but don't add a comment or newline
+  # Designed to feed into Code.Inst constructors, somewhat
+  def toCodeInst(self, params, nonTemporal=0, highBits=0):
+    name = self.name
+    if highBits:
+      name += "_d16_hi"
+    instStr = "%s %s" % (name, (self.formatting % params) )
+    if nonTemporal%2==1:
+      instStr += " glc"
+    if nonTemporal//2==1:
+      instStr += " slc"
+    line = "%-50s" % (instStr)
+    return line
+
+
+  def __str__(self):
+    return self.name
